Move threshold calc to separate file for CUDA
CUDA would not compile electron.cpp if it included the eigen 3.3.4-4
headers. I believe it is a bug in eigen. I tried upgrading eigen to
the latest version, but then lsq would fail to compile. Thus, I
just moved the thresholding function into a separate file.

CUDA can now compile and run the electron code.
