+++ stempy/electron.h
@@ -5,12 +5,6 @@
 
 namespace stempy {
 
-std::pair<double, double> calculateThresholds(std::vector<Block>& blocks,
-                                              Image<double>& darkreference,
-                                              int numberOfSamples = 20,
-                                              int backgroundThresholdNSigma = 4,
-                                              int xRayThresholdNSigma = 10);
-
 std::vector<std::vector<std::pair<int, int>>> electronCount(
   std::vector<Block>& blocks, int scanRows, int scanColumns,
   Image<double>& darkreference, int numberOfSamples = 20,
